Extract `parentNeedsParentheses`
