+++ src/language-js/parentheses/needs-parentheses.js
@@ -1,18 +1,13 @@
-import isNonEmptyArray from "../../utilities/is-non-empty-array.js";
 import {
   createTypeCheckFunction,
   getFunctionParameters,
-  getLeftSidePathName,
   getPrecedence,
   hasComment,
-  hasNakedLeftSide,
-  hasNode,
   isBitwiseOperator,
   isNullishCoalescing,
   isObjectProperty,
   shouldFlatten,
   startsWithNoLookaheadToken,
-  stripChainElementWrappers,
 } from "../utilities/index.js";
 import {
   isArrayExpression,
@@ -27,9 +22,9 @@ import {
   isReturnOrThrowStatement,
   isUnionType,
 } from "../utilities/node-types.js";
-import { returnArgumentHasLeadingComment } from "../utilities/return-statement-has-leading-comment.js";
 import { shouldAddParenthesesToChainElement } from "./chain-expression.js";
 import { shouldAddParenthesesToIdentifier } from "./identifier.js";
+import { parentNeedsParentheses } from "./parent-needs-parentheses.js";
 
 /**
  * @import AstPath from "../../common/ast-path.js"
@@ -102,106 +97,13 @@ function needsParentheses(path, options) {
     }
   }
 
-  switch (parent.type) {
-    case "ReturnStatement":
-    case "ThrowStatement":
-      if (willReturnOrThrowStatementBreak(path, options)) {
-        return false;
-      }
-      break;
-
-    case "ParenthesizedExpression":
-      return false;
-    case "ClassDeclaration":
-    case "ClassExpression":
-      // Add parens around the extends clause of a class. It is needed for almost
-      // all expressions.
-      if (key === "superClass") {
-        const superClass = stripChainElementWrappers(node);
-        if (
-          superClass.type === "ArrowFunctionExpression" ||
-          superClass.type === "AssignmentExpression" ||
-          superClass.type === "AwaitExpression" ||
-          superClass.type === "BinaryExpression" ||
-          superClass.type === "ConditionalExpression" ||
-          superClass.type === "LogicalExpression" ||
-          superClass.type === "NewExpression" ||
-          superClass.type === "ObjectExpression" ||
-          superClass.type === "SequenceExpression" ||
-          superClass.type === "TaggedTemplateExpression" ||
-          superClass.type === "UnaryExpression" ||
-          superClass.type === "UpdateExpression" ||
-          superClass.type === "YieldExpression" ||
-          (superClass.type === "ClassExpression" &&
-            isNonEmptyArray(superClass.decorators))
-        ) {
-          return true;
-        }
-      }
-      break;
-
-    case "ExportDefaultDeclaration":
-      // `export default function` or `export default class` can't be followed by
-      // anything after. So an expression like `export default (function(){}).toString()`
-      // needs to be followed by a parentheses
-      if (shouldWrapFunctionForExportDefault(path, options)) {
-        return true;
-      }
-      break;
-
-    case "Decorator":
-      if (
-        key === "expression" &&
-        !canDecoratorExpressionUnparenthesized(node)
-      ) {
-        return true;
-      }
-      break;
-
-    case "TypeAnnotation":
-      if (
-        path.match(
-          undefined,
-          undefined,
-          (node, key) =>
-            key === "returnType" && node.type === "ArrowFunctionExpression",
-        ) &&
-        includesFunctionTypeInObjectType(node)
-      ) {
-        return true;
-      }
-      break;
-
-    // A user typing `!foo instanceof Bar` probably intended
-    // `!(foo instanceof Bar)`, so format to `(!foo) instance Bar` to what is
-    // really happening
-    case "BinaryExpression":
-      if (
-        key === "left" &&
-        (parent.operator === "in" || parent.operator === "instanceof") &&
-        node.type === "UnaryExpression"
-      ) {
-        return true;
-      }
-      break;
-
-    case "VariableDeclarator":
-      // Legacy syntax
-      // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Errors/Invalid_for-in_initializer
-      // `for (var a = 1 in b);`
-      if (
-        key === "init" &&
-        path.match(
-          undefined,
-          undefined,
-          (node, key) =>
-            key === "declarations" && node.type === "VariableDeclaration",
-          (node, key) => key === "left" && node.type === "ForInStatement",
-        )
-      ) {
-        return true;
-      }
-      break;
+  const parentCheckResult = parentNeedsParentheses(
+    path,
+    options,
+    needsParentheses,
+  );
+  if (typeof parentCheckResult === "boolean") {
+    return parentCheckResult;
   }
 
   switch (node.type) {
@@ -1027,15 +929,6 @@ function isPathInForStatementInitializer(path) {
   return false;
 }
 
-function includesFunctionTypeInObjectType(node) {
-  return hasNode(
-    node,
-    (node) =>
-      node.type === "ObjectTypeAnnotation" &&
-      hasNode(node, (node) => node.type === "FunctionTypeAnnotation"),
-  );
-}
-
 function endsWithRightBracket(node) {
   return isObjectExpression(node);
 }
@@ -1074,96 +967,6 @@ function isFollowedByRightBracket(path) {
       }
       break;
   }
-  return false;
-}
-
-/**
- * @param {AstPath} path
- * @returns {boolean}
- */
-function shouldWrapFunctionForExportDefault(path, options) {
-  const { node, parent } = path;
-
-  if (node.type === "FunctionExpression" || node.type === "ClassExpression") {
-    return (
-      parent.type === "ExportDefaultDeclaration" ||
-      // in some cases the function is already wrapped
-      // (e.g. `export default (function() {})();`)
-      // in this case we don't need to add extra parens
-      !needsParentheses(path, options)
-    );
-  }
-
-  if (
-    !hasNakedLeftSide(node) ||
-    (parent.type !== "ExportDefaultDeclaration" &&
-      needsParentheses(path, options))
-  ) {
-    return false;
-  }
-
-  return path.call(
-    () => shouldWrapFunctionForExportDefault(path, options),
-    ...getLeftSidePathName(node),
-  );
-}
-
-function isDecoratorMemberExpression(node) {
-  if (node.type === "Identifier") {
-    return true;
-  }
-
-  if (isMemberExpression(node)) {
-    return (
-      !node.computed &&
-      // @ts-expect-error -- doesn't exists on `MemberExpression`
-      !node.optional &&
-      node.property.type === "Identifier" &&
-      isDecoratorMemberExpression(node.object)
-    );
-  }
-
-  return false;
-}
-
-// Based on babel implementation
-// https://github.com/nicolo-ribaudo/babel/blob/c4b88a4e5005364255f7e964fe324cf7bfdfb019/packages/babel-generator/src/node/index.ts#L111
-function canDecoratorExpressionUnparenthesized(node) {
-  if (node.type === "ChainExpression") {
-    node = node.expression;
-  }
-
-  return (
-    isDecoratorMemberExpression(node) ||
-    (isCallExpression(node) &&
-      // @ts-expect-error -- doesn't exists on `CallExpression`
-      !node.optional &&
-      isDecoratorMemberExpression(node.callee))
-  );
-}
-
-function willReturnOrThrowStatementBreak(path, options) {
-  const { key, parent } = path;
-  if (!(key === "argument" && isReturnOrThrowStatement(parent))) {
-    return false;
-  }
-
-  /*
-  When `ReturnStatement` or `ThrowStatement` breaks, parentheses will be added around it's argument.
-  So don't need add parentheses again.
-  But we can't know how the argument printed, so only matches cases that will break for sure
-  */
-
-  const { node } = path;
-
-  if (
-    (node.type === "SequenceExpression" ||
-      node.type === "AssignmentExpression") &&
-    returnArgumentHasLeadingComment(node, options)
-  ) {
-    return true;
-  }
-
   return false;
 }
